+++ src/models/vds_software.rs
@@ -17,7 +17,7 @@ use crate::models;
 pub struct VdsSoftware {
     /// ID ПО из маркетплейса.
     #[serde(rename = "id", skip_serializing_if = "Option::is_none")]
-    pub id:   Option<f64>,
+    pub id:   Option<i64>,
     /// Название ПО из маркетплейса.
     #[serde(rename = "name", skip_serializing_if = "Option::is_none")]
     pub name: Option<String>
